+++ include/gaia/ecs/system_base.h
@@ -20,7 +20,7 @@ namespace gaia {
 		class BaseSystemManager;
 
 #if GAIA_PROFILER_CPU
-		constexpr size_t MaxSystemNameLength = 64;
+		constexpr uint32_t MaxSystemNameLength = 64;
 #endif
 
 		class BaseSystem {
@@ -278,9 +278,9 @@ namespace gaia {
 
 		private:
 			void SortSystems() {
-				for (size_t l = 0; l < m_systems.size() - 1; l++) {
+				for (uint32_t l = 0; l < m_systems.size() - 1; l++) {
 					auto min = l;
-					for (size_t p = l + 1; p < m_systems.size(); p++) {
+					for (uint32_t p = l + 1; p < m_systems.size(); p++) {
 						const auto* sl = m_systems[l];
 						const auto* pl = m_systems[p];
 						if (sl->DependsOn(pl))
